src/pipelines/shared/requirements.txt
minio==7.2.15
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
ipython
pyspark== 3.5.4

